rename RAW2PEAK to CENTROIDING
