+++ src/openms/includes.cmake
@@ -39,7 +39,7 @@ include(source/FILTERING/SMOOTHING/sources.cmake)
 include(source/FILTERING/NOISEESTIMATION/sources.cmake)
 include(source/FILTERING/ID/sources.cmake)
 include(source/TRANSFORMATIONS/FEATUREFINDER/sources.cmake)
-include(source/TRANSFORMATIONS/RAW2PEAK/sources.cmake)
+include(source/TRANSFORMATIONS/CENTROIDING/sources.cmake)
 include(source/COMPARISON/CLUSTERING/sources.cmake)
 include(source/COMPARISON/SPECTRA/sources.cmake)
 include(source/QC/sources.cmake)
@@ -95,7 +95,7 @@ include(include/OpenMS/FILTERING/NOISEESTIMATION/sources.cmake)
 include(include/OpenMS/FILTERING/SMOOTHING/sources.cmake)
 include(include/OpenMS/FILTERING/TRANSFORMERS/sources.cmake)
 include(include/OpenMS/TRANSFORMATIONS/FEATUREFINDER/sources.cmake)
-include(include/OpenMS/TRANSFORMATIONS/RAW2PEAK/sources.cmake)
+include(include/OpenMS/TRANSFORMATIONS/CENTROIDING/sources.cmake)
 include(include/OpenMS/QC/sources.cmake)
 if(NOT DISABLE_OPENSWATH)
   include(include/OpenMS/ANALYSIS/OPENSWATH/sources.cmake)
