+++ src/openms/source/TRANSFORMATIONS/CENTROIDING/PeakPickerIterative.cpp
@@ -1,0 +1,16 @@
+// Copyright (c) 2002-present, The OpenMS Team -- EKU Tuebingen, ETH Zurich, and FU Berlin
+// SPDX-License-Identifier: BSD-3-Clause
+//
+// --------------------------------------------------------------------------
+// $Maintainer: Hannes Roest $
+// $Authors: Hannes Roest $
+// --------------------------------------------------------------------------
+
+#include <OpenMS/TRANSFORMATIONS/CENTROIDING/PeakPickerIterative.h>
+
+using namespace std;
+
+namespace OpenMS
+{
+
+}
