+++ src/tests/class_tests/openms/source/PeakPickerIterative_test.cpp
@@ -9,7 +9,7 @@
 #include <OpenMS/CONCEPT/ClassTest.h>
 
 ///////////////////////////
-#include <OpenMS/TRANSFORMATIONS/RAW2PEAK/PeakPickerIterative.h>
+#include <OpenMS/TRANSFORMATIONS/CENTROIDING/PeakPickerIterative.h>
 ///////////////////////////
 
 using namespace OpenMS;
